+++ src/collector/CollectorWrapper.test.ts
@@ -48,7 +48,7 @@ describe('CollectorWrapper', () => {
     ) {
       // We are testing the side effects of the constructor, so we wrap
       // it here to avoid eslint error. We will not disable this rule which as great benefits, but not here.
-      return new CollectorWrapper(options, global.window, sessionIdHandler, testNameHandler)
+      return new CollectorWrapper({ ...options, window: global.window }, sessionIdHandler, testNameHandler)
     }
 
     describe('when debug option is set to true', () => {
@@ -375,8 +375,7 @@ describe('CollectorWrapper', () => {
   describe('identifySession', () => {
     it('delegates session trait to handler', () => {
       const collectorWrapper = new CollectorWrapper(
-        completeOptions({ debug: true }),
-        global.window,
+        completeOptions({ debug: true, window: global.window }),
         new MemorySessionIdHandler(uuidv4, 1000),
         new SessionStorageTestNameHandler(),
       )
@@ -387,8 +386,7 @@ describe('CollectorWrapper', () => {
 
     it('prevents bad format of session trait value', () => {
       const collectorWrapper = new CollectorWrapper(
-        completeOptions({ debug: true }),
-        global.window,
+        completeOptions({ debug: true, window: global.window }),
         new MemorySessionIdHandler(uuidv4, 1000),
         new SessionStorageTestNameHandler(),
       )
@@ -408,8 +406,8 @@ describe('CollectorWrapper', () => {
         completeOptions({
           debug: true,
           sessionsPercentageKept: 100,
+          window: global.window,
         }),
-        global.window,
         new MemorySessionIdHandler(uuidv4, 1000),
         new SessionStorageTestNameHandler(),
       )
@@ -423,8 +421,8 @@ describe('CollectorWrapper', () => {
         completeOptions({
           debug: true,
           sessionsPercentageKept: 0,
+          window: global.window,
         }),
-        global.window,
         new MemorySessionIdHandler(uuidv4, 1000),
         new SessionStorageTestNameHandler(),
       )
@@ -441,8 +439,8 @@ describe('CollectorWrapper', () => {
           completeOptions({
             debug: true,
             sessionsPercentageKept,
+            window: global.window,
           }),
-          global.window,
           memorySessionIdHandler,
           new SessionStorageTestNameHandler(),
         )
@@ -469,8 +467,8 @@ describe('CollectorWrapper', () => {
           completeOptions({
             debug: true,
             sessionsPercentageKept,
+            window: global.window,
           }),
-          global.window,
           new MemorySessionIdHandler(uuidv4, 1000),
           new SessionStorageTestNameHandler(),
         )
@@ -505,8 +503,8 @@ describe('CollectorWrapper', () => {
         completeOptions({
           debug: true,
           rejectSession: () => true,
+          window: global.window,
         }),
-        global.window,
         new MemorySessionIdHandler(uuidv4, 1000),
         new SessionStorageTestNameHandler(),
       )
@@ -520,8 +518,8 @@ describe('CollectorWrapper', () => {
         completeOptions({
           debug: true,
           rejectSession: () => false,
+          window: global.window,
         }),
-        global.window,
         new MemorySessionIdHandler(uuidv4, 1000),
         new SessionStorageTestNameHandler(),
       )