+++ src/session-id-handler/CookieSessionIdHandler.ts
@@ -5,16 +5,20 @@ const GRAVITY_SESSION_ID_COOKIE_KEY = 'gravity_session_id'
 const GRAVITY_SESSION_TIMEOUT_COOKIE_KEY = 'gravity_session_timeout'
 
 export default class CookieSessionIdHandler extends BaseSessionIdHandler implements ISessionIdHandler {
+  public constructor(makeSessionId: () => string, sessionDuration: number, private readonly win: typeof window) {
+    super(makeSessionId, sessionDuration)
+  }
+
   protected getSessionId(): string | undefined {
-    return readCookie(GRAVITY_SESSION_ID_COOKIE_KEY)
+    return readCookie(GRAVITY_SESSION_ID_COOKIE_KEY, this.win)
   }
 
   protected setSessionId(sessionId: string): void {
-    setCookie(GRAVITY_SESSION_ID_COOKIE_KEY, sessionId)
+    setCookie(GRAVITY_SESSION_ID_COOKIE_KEY, sessionId, this.win)
   }
 
   protected getTimeout(): number {
-    const stored = readCookie(GRAVITY_SESSION_TIMEOUT_COOKIE_KEY)
+    const stored = readCookie(GRAVITY_SESSION_TIMEOUT_COOKIE_KEY, this.win)
     return stored !== undefined ? parseInt(stored) : new Date().getTime() - 1
   }