+++ src/session-id-handler/SessionStorageSessionIdHandler.ts
@@ -4,21 +4,25 @@ const GRAVITY_SESSION_STORAGE_KEY_SESSION_ID = 'gravity-session-id'
 const GRAVITY_SESSION_STORAGE_KEY_TIMEOUT = 'gravity-session-timeout'
 
 export default class SessionStorageSessionIdHandler extends BaseSessionIdHandler implements ISessionIdHandler {
+  public constructor(makeSessionId: () => string, sessionDuration: number, private readonly win: typeof window) {
+    super(makeSessionId, sessionDuration)
+  }
+
   protected getSessionId(): string | undefined {
-    return window.sessionStorage.getItem(GRAVITY_SESSION_STORAGE_KEY_SESSION_ID) ?? undefined
+    return this.win.sessionStorage.getItem(GRAVITY_SESSION_STORAGE_KEY_SESSION_ID) ?? undefined
   }
 
   protected setSessionId(sessionId: string): void {
-    window.sessionStorage.setItem(GRAVITY_SESSION_STORAGE_KEY_SESSION_ID, sessionId)
+    this.win.sessionStorage.setItem(GRAVITY_SESSION_STORAGE_KEY_SESSION_ID, sessionId)
   }
 
   protected getTimeout(): number {
-    const stored = window.sessionStorage.getItem(GRAVITY_SESSION_STORAGE_KEY_TIMEOUT)
+    const stored = this.win.sessionStorage.getItem(GRAVITY_SESSION_STORAGE_KEY_TIMEOUT)
 
     return stored !== null ? parseInt(stored) : new Date().getTime() - 1
   }
 
   protected setTimeout(timeout: number) {
-    window.sessionStorage.setItem(GRAVITY_SESSION_STORAGE_KEY_TIMEOUT, timeout.toString())
+    this.win.sessionStorage.setItem(GRAVITY_SESSION_STORAGE_KEY_TIMEOUT, timeout.toString())
   }
 }