+++ src/types.ts
@@ -191,7 +191,12 @@ export interface CollectorOptions {
    */
   originsToRecord?: string[]
   recordRequestsFor?: string[]
+  window?: typeof window
   enabledListeners?: Listener[]
+}
+
+export type CollectorOptionsWithWindow = CollectorOptions & {
+  window: typeof window
 }
 
 export enum Listener {